roles and permissions

+++ routes/web.php
@@ -3,6 +3,9 @@
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UsersController;
+use App\Http\Controllers\RoleController;
+use App\Http\Controllers\PermissionController;
+
 
 Route::get('/', function () {
     return view('dashboard');
@@ -14,12 +17,26 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
     Route::group(['prefix'=>'users'], function(){
-        Route::get('/',                             [UsersController::class, 'index'])->name('users.index');
-        Route::get('/create',                       [UsersController::class, 'create'])->name('users.create');
-        Route::post('/create',                      [UsersController::class, 'store'])->name('users.store');
-        Route::get('/{user}',                       [UsersController::class, 'show'])->name('users.show');
-        Route::put('/{user}',                       [UsersController::class, 'update'])->name('users.update');
-        Route::delete('/{user}',                    [UsersController::class, 'destroy'])->name('users.destroy');
+        Route::get('/',                             [UsersController::class, 'index'])->name('users.index')->middleware('checkPermissions:read_users');
+        Route::get('/create',                       [UsersController::class, 'create'])->name('users.create')->middleware('checkPermissions:write_users');
+        Route::post('/create',                      [UsersController::class, 'store'])->name('users.store')->middleware('checkPermissions:write_users');
+        Route::get('/{user}',                       [UsersController::class, 'show'])->name('users.show')->middleware('checkPermissions:read_users');
+        Route::put('/{user}',                       [UsersController::class, 'update'])->name('users.update')->middleware('checkPermissions:edit_users');
+        Route::delete('/{user}',                    [UsersController::class, 'destroy'])->name('users.destroy')->middleware('checkPermissions:delete_users');
+    });
+
+    Route::group(['prefix'=>'roles'], function(){
+        Route::get('/',                 [RoleController::class, 'index'])->name('roles.index')->middleware('checkPermissions:read_roles');
+        Route::get('/create',           [RoleController::class, 'create'])->name('roles.create')->middleware('checkPermissions:write_roles');
+        Route::post('/create',          [RoleController::class, 'store'])->name('roles.store')->middleware('checkPermissions:write_roles');
+        Route::get('/{role}',           [RoleController::class, 'show'])->name('roles.show')->middleware('checkPermissions:read_roles');
+        Route::put('/{role}',           [RoleController::class, 'update'])->name('roles.update')->middleware('checkPermissions:edit_roles');
+        Route::delete('/{role}',        [RoleController::class, 'destroy'])->name('roles.destroy')->middleware('checkPermissions:delete_roles');
+    });
+    
+    Route::group(['prefix'=>'permissions'], function(){
+        Route::get('/',                 [PermissionController::class, 'index'])->name('permissions.index')->middleware('checkPermissions:read_permissions');
+        Route::get('/{permission}',     [PermissionController::class, 'show'])->name('permissions.show')->middleware('checkPermissions:read_permissions');
     });
 });
